+++ src/components/card/styled.tsx
